+++ sash-utils/sash-utils.sh
@@ -1,6 +1,7 @@
 #!/usr/bin/env bash
 
 _SASH_UTILS_MY_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+source "$_SASH_UTILS_MY_DIR/err-stack.sh"
 source "$_SASH_UTILS_MY_DIR/colors.sh"
 source "$_SASH_UTILS_MY_DIR/choices.sh"
 source "$_SASH_UTILS_MY_DIR/dirs.sh"