+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Don't source `sash-parse.sh` before plugins load as plugins can overwrite that.
 * Move `sash-parse.sh` into a `sash-libs` folder.
+* Create `sash-utils` for basic utility script for sash directly.
+* Use `trap` in order to prevent IFS mixups.
 
 ## 1.1.1 (November 5th, 2018)